requirements.txt
binance
pandas
pandas_ta
pyyaml
python-dotenv
